Refactor database setup: updated mongoose configuration and removed outdated models
