--- lib/darabase/mongoose.ts
@@ -1,39 +1,0 @@
-import mongoose, { Mongoose } from "mongoose";
-
-const MONGODB_URL = process.env.MONGODB_URL;
-
-if (!MONGODB_URL) {
-  throw new Error("Environment variable MONGODB_URL is not defined");
-}
-
-interface MongooseConnection {
-  conn: Mongoose | null;
-  promise: Promise<Mongoose> | null;
-}
-
-declare global {
-  var mongoose: MongooseConnection | undefined;
-}
-
-let cached: MongooseConnection = global.mongoose || {
-  conn: null,
-  promise: null,
-};
-
-global.mongoose = cached;
-
-export const connectionToDatabase = async (): Promise<Mongoose> => {
-  if (cached.conn) {
-    return cached.conn;
-  }
-
-  if (!cached.promise) {
-    cached.promise = mongoose.connect(MONGODB_URL, {
-      dbName: "AraOkan",
-      bufferCommands: false,
-    });
-  }
-
-  cached.conn = await cached.promise;
-  return cached.conn;
-};
