+++ lib/database/mongoose.ts
@@ -1,41 +1,40 @@
+"use server";
 import mongoose, { Mongoose } from "mongoose";
 
 const MONGODB_URL = process.env.MONGODB_URL;
+
+if (!MONGODB_URL) {
+  throw new Error("Environment variable MONGODB_URL is not defined");
+}
 
 interface MongooseConnection {
   conn: Mongoose | null;
   promise: Promise<Mongoose> | null;
 }
 
-// Use `let` or `const` appropriately
-let cached: MongooseConnection = (
-  global as unknown as { mongoose?: MongooseConnection }
-).mongoose || {
+declare global {
+  var mongoose: MongooseConnection | undefined;
+}
+
+let cached: MongooseConnection = global.mongoose || {
   conn: null,
   promise: null,
 };
 
-if (!cached.conn) {
-  cached = (global as unknown as { mongoose: MongooseConnection }).mongoose = {
-    conn: null,
-    promise: null,
-  };
-}
+global.mongoose = cached;
 
-export const connectToDatabase = async (): Promise<Mongoose> => {
-  if (cached.conn) return cached.conn;
+export const connectionToDatabase = async (): Promise<Mongoose> => {
+  if (cached.conn) {
+    return cached.conn;
+  }
 
-  if (!MONGODB_URL) throw new Error("Missing MONGODB_URL");
-
-  const promise =
-    cached.promise ||
-    mongoose.connect(MONGODB_URL, {
+  if (!cached.promise) {
+    cached.promise = mongoose.connect(MONGODB_URL, {
       dbName: "AraOkan",
       bufferCommands: false,
     });
+  }
 
-  cached.promise = promise;
-  cached.conn = await promise;
-
+  cached.conn = await cached.promise;
   return cached.conn;
 };
